+++ openghg/store/_footprints.py
@@ -24,7 +24,7 @@ class Footprints(BaseStore):
     _uuid = "62db5bdf-c88d-4e56-97f4-40336d37f18c"
     _metakey = f"{_root}/uuid/{_uuid}/metastore"
 
-    def read_data(self, binary_data: bytes, metadata: dict, file_metadata: dict) -> dict | None:
+    def read_data(self, binary_data: bytes, metadata: dict, file_metadata: dict) -> list[dict] | None:
         """Ready a footprint from binary data
 
         Args:
@@ -186,10 +186,13 @@ class Footprints(BaseStore):
 
     def read_file(
         self,
-        filepath: list | str | Path,
-        site: str,
         domain: str,
         model: str,
+        filepath: list | str | Path,
+        site: str | None = None,
+        satellite: str | None = None,
+        obs_region: str | None = None,
+        selection: str | None = None,
         inlet: str | None = None,
         height: str | None = None,
         met_model: str | None = None,
@@ -213,7 +216,7 @@ class Footprints(BaseStore):
         compressor: Any | None = None,
         filters: Any | None = None,
         optional_metadata: dict | None = None,
-    ) -> dict:
+    ) -> list[dict]:
         """Reads footprints data files and returns the UUIDS of the Datasources
         the processed data has been assigned to
 
@@ -221,6 +224,8 @@ class Footprints(BaseStore):
             filepath: Path(s) of file(s) to standardise
             site: Site name
             domain: Domain of footprints
+            satellite: Satellite name
+            obs_region: The geographic region covered by the data ("BRAZIL", "INDIA", "UK").
             model: Model used to create footprint (e.g. NAME or FLEXPART)
             inlet: Height above ground level in metres. Format 'NUMUNIT' e.g. "10m"
             height: Alias for inlet. One of height or inlet MUST be included.
@@ -283,7 +288,14 @@ class Footprints(BaseStore):
             )
             time_resolved = high_time_resolution
 
-        site = clean_string(site)
+        if site is not None:
+            site = clean_string(site)
+        elif satellite is not None and obs_region is not None:
+            satellite = clean_string(satellite)
+            obs_region = clean_string(obs_region)
+        else:
+            raise ValueError("Please pass either site or satellite and obs_region values")
+
         network = clean_string(network)
         domain = clean_string(domain)
 
@@ -356,12 +368,12 @@ class Footprints(BaseStore):
         _, unseen_hashes = self.check_hashes(filepaths=filepath, force=force)
 
         if not unseen_hashes:
-            return {}
+            return [{}]
 
         filepath = list(unseen_hashes.values())
 
         if not filepath:
-            return {}
+            return [{}]
 
         # Define parameters to pass to the parser function and remaining keys
         parser_input_parameters, additional_input_parameters = split_function_inputs(
@@ -371,7 +383,7 @@ class Footprints(BaseStore):
         footprint_data = parser_fn(**parser_input_parameters)
 
         chunks = self.check_chunks(
-            ds=list(footprint_data.values())[0]["data"],
+            ds=footprint_data[0].data,
             chunks=chunks,
             high_spatial_resolution=high_spatial_resolution,
             time_resolved=time_resolved,
@@ -383,13 +395,11 @@ class Footprints(BaseStore):
         # Checking against expected format for footprints
         # Based on configuration (some user defined, some inferred)
         # Also check for alignment of domain coordinates
-        for split_data in footprint_data.values():
+        for mdd in footprint_data:
+            mdd.data = mdd.data.chunk(chunks)
 
-            split_data["data"] = split_data["data"].chunk(chunks)
-
-            fp_data = split_data["data"]
             Footprints.validate_data(
-                fp_data,
+                mdd.data,
                 high_spatial_resolution=high_spatial_resolution,
                 time_resolved=time_resolved,
                 short_lifetime=short_lifetime,